(maint) Refactor utilities, the separation was sort of arbitrary
