+++ bin/subcommands/version
@@ -3,4 +3,4 @@ load File.expand_path(File.join(File.dirname(__FILE__), "..", "..", "lib", "pupp
 require 'puppet_docker_tools/utilities'
 
 directory = ARGV[0] || '.'
-puts PuppetDockerTools::Utilities::Dockerfile.get_value_from_env('version', directory: directory)
+puts PuppetDockerTools::Utilities.get_value_from_env('version', directory: directory)
